Consolidate gpt-oss MXFP4 patch tests into a single file
Merge the post-load swizzle walker and pre-load gate tests into a
single behavior-named module covering both halves of patch_gpt_oss():

- Post-load native swizzle walker: zero-placeholder skip vs raw-block
  raise, meta-scales handling, per-projection cache skip including the
  gate-up-cached-but-down-raw partial case.
- Pre-load wrapper: use_kernels=True keeps native CPU MXFP4, default
  call forces dequantize, positional argument compatibility,
  use_kernels forwarded to upstream, detection-failure warns.
